==> notebooks/01_ingesta.ipynb <==
# Importar librerías
from pyspark.sql import SparkSession
from pyspark.sql.functions import col, to_date, year, month
import os

spark = (SparkSession.builder
         .appName("SECOP_Ingesta")
         .master("spark://spark-master:7077")
         .config("spark.driver.memory", "2g")
         .config("spark.executor.memory", "2g")
         .config("spark.executor.cores", "1")
         .getOrCreate())


spark.sparkContext.setLogLevel("WARN")
print("Spark master:", spark.sparkContext.master)
print("Spark UI:", spark.sparkContext.uiWebUrl)
print(f"Spark Version: {spark.version}")
print(f"Spark Master: {spark.sparkContext.master}")

from sodapy import Socrata
import json, os
import time

client = Socrata("www.datos.gov.co", None)

out_file = "/opt/spark-data/raw/secop_q4_2025.jsonl"
open(out_file, "w").close()

total = 0

for offset in range(0, 60000, 5000):
    print(f"Consultando offset {offset}...")
    
    results = client.get(
        "jbjy-vk9h",
        query=f"""
        SELECT *
        WHERE fecha_de_firma >= '2025-10-01T00:00:00'
        AND fecha_de_firma <= '2025-12-31T23:59:59'
        LIMIT 5000
        OFFSET {offset}
        """
    )

    print("Recibidos:", len(results))

    with open(out_file, "a", encoding="utf-8") as f:
        for row in results:
            f.write(json.dumps(row, ensure_ascii=False) + "\n")

    total += len(results)
    print("Total acumulado:", total)
    
    time.sleep(1)  # evita throttling agresivo

print("TOTAL FINAL:", total)



import os
size_mb = os.path.getsize("/opt/spark-data/raw/secop_q4_2025.jsonl") / 1024 / 1024
print("Tamaño archivo MB:", round(size_mb, 2))


json_path = "/opt/spark-data/raw/secop_q4_2025.jsonl"
df_raw = spark.read.json(json_path)

print("Total de registros:", df_raw.count())
print("Total de columnas:", len(df_raw.columns))
df_raw.show(5, truncate=False)


df_raw.printSchema()

df_raw.show()

output_path = "/opt/spark-data/processed/secop_bronze_q4_2025"

(df_raw.write
 .mode("overwrite")
 .parquet(output_path))


spark.read.parquet(output_path).count()
